Show a simpler led color picker for crappy led devices [1/3]
some devices like Honor v10 have a led that seems unable
to show all colors in the HSV range but only a few of them,
so in this case we show a different version of the picker
with only available ones.
This is a kang of the awesome Alex's @Mazda-- accents picker
adapted to work like a color picker.

To enable it, overlay:
config_isCrappyLedDevice bool
in the device tree, and add:
isledPicker="true"
to the ColorPickerPreference if the picker is intended to
be used within the led options.